Доработал компонент FavoritesList
