+++ src/pages/favorites-screen/favorites-screen.tsx
@@ -1,34 +1,12 @@
 import { Helmet } from 'react-helmet-async';
 import { OfferType } from '../../types/offer';
-import FavoritesItem from '../../components/favorites-item/favorites-item';
-
-type CityGroupType = {
-  [key: string]: OfferType[];
-}
-
-function groupOffersByCity(items: OfferType[]): CityGroupType {
-
-  const groupedItems = items.reduce((accumulator: CityGroupType, item) => {
-    const cityName = item.city.name;
-
-    if (!accumulator[cityName]) {
-      accumulator[cityName] = [];
-    }
-
-    accumulator[cityName].push(item);
-
-    return accumulator;
-  }, {});
-
-  return groupedItems;
-}
+import FavoritesList from '../../components/favorites-list/favorites-list';
 
 type FavoritesScreenProps = {
   offers: OfferType[];
 }
 
 function FavoritesScreen({offers}: FavoritesScreenProps): JSX.Element {
-  const offersGroupedByCity = groupOffersByCity(offers);
 
   return (
     <main className="page__main page__main--favorites">
@@ -40,10 +18,11 @@ function FavoritesScreen({offers}: FavoritesScreenProps): JSX.Element {
       <div className="page__favorites-container container">
         <section className="favorites">
           <h1 className="favorites__title">Saved listing</h1>
-          <ul className="favorites__list">
-            {/* <FavoritesItem offers={favoriteOffersAmsterdam} /> */}
-            {Object.keys(offersGroupedByCity).map((city) => <FavoritesItem key={city} city={city} offers={offersGroupedByCity[city]}/>)}
-          </ul>
+
+          <FavoritesList
+            offers={offers}
+          />
+
         </section>
       </div>
     </main>
